Refactor role data into shared module
